Remove lib/irc-appservice.js and lib/config/validator.js and app.js
Replaced with appservice-bridge classes. Move cli.js to app.js
