+++ lib/main.js
@@ -2,26 +2,24 @@
 var Promise = require("bluebird");
 var extend = require("extend");
 
-var Cli = require("matrix-appservice-bridge").Cli;
 var AppServiceRegistration = require("matrix-appservice-bridge").AppServiceRegistration;
 var AppService = require("matrix-appservice").AppService;
 
-var ircToMatrix = require("./lib/bridge/irc-to-matrix.js");
-var matrixToIrc = require("./lib/bridge/matrix-to-irc.js");
-var membershiplists = require("./lib/bridge/membershiplists.js");
-var IrcServer = require("./lib/irclib/server.js").IrcServer;
-var ircLib = require("./lib/irclib/irc.js");
-var matrixLib = require("./lib/mxlib/matrix");
-var MatrixUser = require("./lib/models/users").MatrixUser;
-var store = require("./lib/store");
-var stats = require("./lib/config/stats");
-var ident = require("./lib/irclib/ident");
-var names = require("./lib/irclib/names");
-var logging = require("./lib/logging");
+var ircToMatrix = require("./bridge/irc-to-matrix.js");
+var matrixToIrc = require("./bridge/matrix-to-irc.js");
+var membershiplists = require("./bridge/membershiplists.js");
+var IrcServer = require("./irclib/server.js").IrcServer;
+var ircLib = require("./irclib/irc.js");
+var matrixLib = require("./mxlib/matrix");
+var MatrixUser = require("./models/users").MatrixUser;
+var store = require("./store");
+var stats = require("./config/stats");
+var ident = require("./irclib/ident");
+var names = require("./irclib/names");
+var logging = require("./logging");
 var log = logging.get("main");
 
 const DEFAULT_LOCALPART = "appservice-irc";
-const REG_PATH = "appservice-registration-irc.yaml";
 
 var _toServer = function(domain, serverConfig) {
     // set server config defaults
@@ -82,7 +80,7 @@ var _toServer = function(domain, serverConfig) {
     return new IrcServer(domain, extend(true, defaultServerConfig, serverConfig));
 };
 
-var _generateRegistration = Promise.coroutine(function*(reg, config) {
+module.exports.generateRegistration = Promise.coroutine(function*(reg, config) {
     var asToken;
     if (config.appService) {
         console.warn(
@@ -134,7 +132,7 @@ var _generateRegistration = Promise.coroutine(function*(reg, config) {
     return reg;
 });
 
-var _runBridge = Promise.coroutine(function*(port, config, reg) {
+module.exports.runBridge = Promise.coroutine(function*(port, config, reg) {
     if (config.ircService.logging) {
         logging.configure(config.ircService.logging);
         logging.setUncaughtExceptionLogger(log);
@@ -239,38 +237,3 @@ var _runBridge = Promise.coroutine(function*(port, config, reg) {
     };
     appService.listen(port);
 });
-
-
-new Cli({
-    registrationPath: REG_PATH,
-    enableRegistration: true,
-    enableLocalpart: true,
-    bridgeConfig: {
-        affectsRegistration: true,
-        schema: "./lib/config/schema.yml",
-        defaults: {
-            ircService: {
-                ident: {
-                    enabled: false,
-                    port: 113
-                },
-                logging: {
-                    level: "debug",
-                    toConsole: true
-                },
-                statsd: {}
-            }
-        }
-    },
-    generateRegistration: function(reg, callback) {
-        _generateRegistration(reg, this.getConfig()).done(function(completeRegistration) {
-            callback(completeRegistration);
-        });
-    },
-    run: function(port, config, reg) {
-        _runBridge(port, config, reg).catch(function(err) {
-            console.error("Failed to run bridge.");
-            throw err;
-        });
-    }
-}).run();
